fixed a few other UIs
