+++ src/components/Instances/InstanceActions.tsx
@@ -1,52 +1,25 @@
 "use client";
 
-import { InstanceFlavorSelectDialog } from "@/components/Instances/InstanceFlavorSelectDialog";
-import { InstanceVolumeDialog } from "@/components/Instances/InstanceVolumeDialog";
-import { SnapshotDialog } from "@/components/Instances/SnapshotDialog";
 import { ConfirmDeleteDialog } from "@/components/reusable/ConfirmDeleteDialog";
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuPortal,
-  DropdownMenuSeparator,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { FlavorService, InfraService } from "@/lib/requests";
+import { InfraService } from "@/lib/requests";
 import type {
-  IdRequest,
   InstanceDeleteRequest,
   InstanceRebootRequest,
   InstanceStartRequest,
   InstanceStopRequest,
-  ResizeRequest,
 } from "@/types/RequestInterfaces";
-import type { InstanceListItem } from "@/types/ResponseInterfaces";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import {
-  Box,
-  Camera,
-  FileText,
-  GripVertical,
-  HardDrive,
-  Pause,
-  Power,
-  PowerOff,
-  RotateCcw,
-  ShieldAlert,
-  Terminal,
-  Trash2,
-} from "lucide-react";
-import Link from "next/link";
+import type {
+  InstanceListItem,
+  InstanceStatus,
+} from "@/types/ResponseInterfaces";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { Power, PowerOff, RotateCcw, Trash2 } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
 
@@ -54,131 +27,13 @@ export function InstanceActions({
   instance,
   status,
   disabled = false,
-  onViewLogs,
 }: {
   instance: InstanceListItem;
-  status: string;
+  status: InstanceStatus;
   disabled?: boolean;
-  onViewLogs?: (instance: InstanceListItem) => void;
 }) {
   const queryClient = useQueryClient();
   const isDisabled = disabled || status === "BUILD";
-  const [isResizeDialogOpen, setIsResizeDialogOpen] = useState(false);
-  const [isVolumeDialogOpen, setIsVolumeDialogOpen] = useState(false);
-  const [showSnapshotDialog, setShowSnapshotDialog] = useState(false);
-
-  const {
-    data: consoleData,
-    isLoading: isConsoleLoading,
-    refetch: refetchConsole,
-  } = useQuery({
-    queryKey: ["console", instance.id],
-    queryFn: () => InfraService.getConsole({ instance_id: instance.id }),
-    enabled: false,
-    refetchOnWindowFocus: false,
-    retry: 1,
-  });
-
-  const { data: flavors = [] } = useQuery({
-    queryKey: ["flavors"],
-    queryFn: () => FlavorService.list(),
-    staleTime: 5 * 60 * 1000,
-  });
-
-  const consoleUrl = consoleData?.url;
-  const currentFlavorId = flavors?.find((f) => f.name === instance.flavor)?.id;
-
-  const supportedActions = {
-    pause: true,
-    suspend: true,
-    shelve: true,
-    rescue: true,
-    console: true,
-    logs: true,
-    resize: true,
-    volumes: true,
-  };
-
-  const handleViewLogs = () => {
-    if (onViewLogs) {
-      onViewLogs(instance);
-    }
-  };
-
-  const pauseMutation = useMutation({
-    mutationFn: (data: IdRequest) => InfraService.pause(data),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: ["instances"] }),
-        queryClient.invalidateQueries({ queryKey: ["instance", instance.id] }),
-      ]);
-      toast.success("Instance paused successfully");
-    },
-    onError: (err: unknown) =>
-      toast.error(err instanceof Error ? err.message : "Failed to pause"),
-  });
-
-  const suspendMutation = useMutation({
-    mutationFn: (data: IdRequest) => InfraService.suspend(data),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: ["instances"] }),
-        queryClient.invalidateQueries({ queryKey: ["instance", instance.id] }),
-      ]);
-      toast.success("Instance suspended successfully");
-    },
-    onError: (err: unknown) =>
-      toast.error(err instanceof Error ? err.message : "Failed to suspend"),
-  });
-
-  const shelveMutation = useMutation({
-    mutationFn: (data: IdRequest) => InfraService.shelve(data),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: ["instances"] }),
-        queryClient.invalidateQueries({ queryKey: ["instance", instance.id] }),
-      ]);
-      toast.success("Instance shelved successfully");
-    },
-    onError: (err: unknown) =>
-      toast.error(err instanceof Error ? err.message : "Failed to shelve"),
-  });
-
-  const rescueMutation = useMutation({
-    mutationFn: (data: IdRequest) => InfraService.rescue(data),
-    onSuccess: async () => {
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: ["instances"] }),
-        queryClient.invalidateQueries({ queryKey: ["instance", instance.id] }),
-      ]);
-      toast.success("Rescue mode activated");
-    },
-    onError: (err: unknown) =>
-      toast.error(err instanceof Error ? err.message : "Failed to rescue"),
-  });
-
-  const resizeMutation = useMutation({
-    mutationFn: (data: ResizeRequest) => InfraService.resize(data),
-    onSuccess: async () => {
-      toast.success("Instance resize started successfully");
-      await Promise.all([
-        queryClient.invalidateQueries({ queryKey: ["instances"] }),
-        queryClient.invalidateQueries({ queryKey: ["instance", instance.id] }),
-      ]);
-    },
-    onError: (error: Error) => {
-      toast.error(`Failed to resize instance: ${error.message}`);
-    },
-  });
-
-  const handleResize = (flavorId: string) => {
-    if (flavorId) {
-      resizeMutation.mutate({
-        instance_id: instance.id,
-        new_flavor: flavorId,
-      });
-    }
-  };
   const [showDeleteDialog, setShowDeleteDialog] = useState(false);
 
   const deleteMutation = useMutation({
@@ -264,8 +119,6 @@ export function InstanceActions({
   const canStart = status === "SHUTOFF" && !isDisabled;
   const canStop = status === "ACTIVE" && !isDisabled;
   const canReboot = status === "ACTIVE" && !isDisabled;
-  const canSnapshot =
-    !isDisabled && (status === "ACTIVE" || status === "SHUTOFF");
 
   return (
     <div className="flex flex-wrap items-center justify-center gap-2">
@@ -338,158 +191,6 @@ export function InstanceActions({
         </Tooltip>
       )}
 
-      <DropdownMenu>
-        <Tooltip>
-          <TooltipTrigger asChild>
-            <DropdownMenuTrigger asChild>
-              <Button
-                variant="outline"
-                disabled={isDisabled}
-                size="sm"
-                className="group bg-background text-foreground border-border hover:bg-accent hover:text-accent-foreground cursor-pointer rounded-full transition-all duration-200"
-              >
-                <GripVertical className="group-hover:text-accent-foreground mr-1 h-4 w-4 transition-colors duration-200" />
-                <span className="group-hover:text-accent-foreground transition-colors duration-200">
-                  Instance Actions
-                </span>
-              </Button>
-            </DropdownMenuTrigger>
-          </TooltipTrigger>
-          <TooltipContent>Instance actions</TooltipContent>
-        </Tooltip>
-        <DropdownMenuContent align="end">
-          {supportedActions.pause && (
-            <DropdownMenuItem
-              onClick={(e) => {
-                e.stopPropagation();
-                pauseMutation.mutate({ instance_id: instance.id });
-              }}
-              disabled={isDisabled || status !== "ACTIVE"}
-            >
-              <Pause className="mr-2 h-4 w-4" /> Pause
-            </DropdownMenuItem>
-          )}
-          {supportedActions.suspend && (
-            <DropdownMenuItem
-              onClick={(e) => {
-                e.stopPropagation();
-                suspendMutation.mutate({ instance_id: instance.id });
-              }}
-              disabled={isDisabled || status !== "ACTIVE"}
-            >
-              <Pause className="mr-2 h-4 w-4" /> Suspend
-            </DropdownMenuItem>
-          )}
-          {supportedActions.shelve && (
-            <DropdownMenuItem
-              onClick={(e) => {
-                e.stopPropagation();
-                shelveMutation.mutate({ instance_id: instance.id });
-              }}
-              disabled={isDisabled || status !== "ACTIVE"}
-            >
-              <Pause className="mr-2 h-4 w-4" /> Shelve
-            </DropdownMenuItem>
-          )}
-          {supportedActions.rescue && (
-            <DropdownMenuItem
-              onClick={(e) => {
-                e.stopPropagation();
-                rescueMutation.mutate({ instance_id: instance.id });
-              }}
-              disabled={isDisabled || status !== "ACTIVE"}
-            >
-              <ShieldAlert className="mr-2 h-4 w-4" /> Rescue Mode
-            </DropdownMenuItem>
-          )}
-
-          <DropdownMenuSeparator />
-          <DropdownMenuSub>
-            <DropdownMenuSubTrigger
-              disabled={isDisabled}
-              onMouseEnter={() => refetchConsole()}
-            >
-              <Terminal className="mr-2 h-4 w-4" /> Console
-            </DropdownMenuSubTrigger>
-            <DropdownMenuPortal>
-              <DropdownMenuSubContent>
-                {supportedActions.console && (
-                  <Link
-                    href={consoleUrl ?? "#"}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    onClick={(e) => e.stopPropagation()}
-                    passHref
-                  >
-                    <DropdownMenuItem
-                      onClick={(e) => {
-                        if (!consoleUrl) {
-                          e.preventDefault();
-                          e.stopPropagation();
-                          toast.error("Console URL not available");
-                        }
-                      }}
-                      disabled={isDisabled || isConsoleLoading || !consoleUrl}
-                    >
-                      <Terminal className="mr-2 h-4 w-4" />
-                      {isConsoleLoading ? "Loading console..." : "Open Console"}
-                    </DropdownMenuItem>
-                  </Link>
-                )}
-                {supportedActions.logs && (
-                  <DropdownMenuItem
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                      handleViewLogs();
-                    }}
-                    disabled={isDisabled}
-                  >
-                    <FileText className="mr-2 h-4 w-4" /> View Logs
-                  </DropdownMenuItem>
-                )}
-              </DropdownMenuSubContent>
-            </DropdownMenuPortal>
-          </DropdownMenuSub>
-
-          {supportedActions.resize && (
-            <>
-              <DropdownMenuSeparator />
-              <DropdownMenuItem
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setIsResizeDialogOpen(true);
-                }}
-                disabled={!["ACTIVE", "SHUTOFF"].includes(status)}
-              >
-                <Box className="mr-2 h-4 w-4" /> Resize Instance
-              </DropdownMenuItem>
-              {canSnapshot && (
-                <DropdownMenuItem
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    setShowSnapshotDialog(true);
-                  }}
-                  disabled={isDisabled}
-                >
-                  <Camera className="mr-2 h-4 w-4" /> Create Snapshot
-                </DropdownMenuItem>
-              )}
-              <DropdownMenuItem
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setIsVolumeDialogOpen(true);
-                }}
-                disabled={isDisabled}
-              >
-                <HardDrive className="mr-2 h-4 w-4" />{" "}
-                {instance.has_volume ? "Detach Volume" : "Attach Volume"}
-              </DropdownMenuItem>
-            </>
-          )}
-        </DropdownMenuContent>
-      </DropdownMenu>
-
       <Tooltip>
         <TooltipTrigger asChild>
           <Button
@@ -522,31 +223,6 @@ export function InstanceActions({
           }
           confirming={deleteMutation.isPending}
           onConfirm={() => deleteMutation.mutate({ server_id: instance.id })}
-        />
-      </div>
-      <div onClick={(e) => e.stopPropagation()}>
-        <InstanceVolumeDialog
-          open={isVolumeDialogOpen}
-          onOpenChange={setIsVolumeDialogOpen}
-          instanceId={instance.id}
-          hasVolume={instance.has_volume}
-        />
-      </div>
-      <div onClick={(e) => e.stopPropagation()}>
-        <InstanceFlavorSelectDialog
-          open={isResizeDialogOpen}
-          onOpenChange={setIsResizeDialogOpen}
-          currentFlavorId={currentFlavorId}
-          onSelect={handleResize}
-        />
-      </div>
-      <div onClick={(e) => e.stopPropagation()}>
-        <SnapshotDialog
-          open={showSnapshotDialog}
-          onOpenChange={setShowSnapshotDialog}
-          instanceId={instance.id}
-          instanceName={instance.instance_name}
-          disabled={isDisabled}
         />
       </div>
     </div>
